add components

+++ lib/config/numbers/app_numbers_data.dart
@@ -6,17 +6,17 @@ class AppNumbersData {
 
   // BREAKPOINTS
 
-  int get mobileBreak => 375;
-  int get tabletBreak => 768;
-  int get desktopBreak => 1024;
+  double get mobileBreak => 375;
+  double get tabletBreak => 768;
+  double get desktopBreak => 1024;
 
   // BUTTONS HEIGHT
 
-  int get buttonXS => 24;
-  int get buttonS => 32;
-  int get buttonM => 36;
-  int get buttonL => 40;
-  int get buttonXL => 48;
+  double get buttonXS => 24;
+  double get buttonS => 32;
+  double get buttonM => 36;
+  double get buttonL => 40;
+  double get buttonXL => 48;
 
   BorderRadiusData get brRadius => const BorderRadiusData();
   SpacingsData get spacings => const SpacingsData();
