+++ lib/config/numbers/spacings_data.dart
@@ -1,18 +1,18 @@
 class SpacingsData {
   const SpacingsData();
 
-  int get x_5 => 2;
-  int get x1 => 4;
-  int get x1_5 => 6;
-  int get x2 => 8;
-  int get x2_5 => 10;
-  int get x3 => 12;
-  int get x4 => 16;
-  int get x5 => 20;
-  int get x6 => 24;
-  int get x7 => 28;
-  int get x8 => 32;
-  int get x10 => 40;
-  int get x12 => 48;
-  int get x16 => 64;
+  double get x_5 => 2;
+  double get x1 => 4;
+  double get x1_5 => 6;
+  double get x2 => 8;
+  double get x2_5 => 10;
+  double get x3 => 12;
+  double get x4 => 16;
+  double get x5 => 20;
+  double get x6 => 24;
+  double get x7 => 28;
+  double get x8 => 32;
+  double get x10 => 40;
+  double get x12 => 48;
+  double get x16 => 64;
 }
